Adding optional OpenTracing integration.
[OpenTracing](https://opentracing.io/) provides the ability for
distributed tracing independent of vendor or downstream consumer.

This commit optionally adds an OpenTracing span in the receiver methods. The
spans are defined as `django-auditlog.{action}.{model_name}`, where
`action` is either `create`, `update`, or `delete`. This allows for
analysis on the performance impacted created through audit logging.

To install it, include the extra "opentracing"
Ex: `pip install django-auditlog[opentracing]`