Add support for using DB column in diff.
Added a new field `db_fields` to `register`. Any field specified here
will use the DB column name, instead of the field name. If a custom DB
column was specified when defining the field, that will be used.

Relations are not followed, leading to significant performance gains
when a relational field is specified in `db_fields`. For example, if a
field called `"relation"` exists and is specified in `db_fields`,
`relation_id` is used (or the custom column name) and `relation` is
excluded.