Adding Foriegn Key Only Fields
When building the changes JSON for a LogEntry, following
relations (e.g. `ForeigKey` fields) can resutl in a
significant performance penalty. This commit adds support for
only logging to foreign key value, by specifying
fields to log in this manner wiht a `fk_only_fields`
when registering a model with the audit log.