Implement ``mask_fields`` in the serialized data
A ``MaskedDictionary`` object was created to take the serialized data
and recursively inspect all fields for matches to fields that should be
masked. A recursive implementation was chosen because it will handle any
nested relationships that may be serialized. An added benefit of this
approach is the ability to specify keys within a serialized JSONField that
may need to be masked.