Moved masking to model and added field restriction
Field masking was simplified by removing the recursive approach for
masking nested fields. The simplfied implementation was moved to the
model to be consistent with the addition of a field restriction option.
The fields that are included in the serailized data can now be
restricted such that they honor the ``include_fields`` and
``exclude_fields`` kwargs. This behavior can be enabled by adding
``serialize_auditlog_fields_only=True`` during model registration.