Type fields before sending to serializer
The Django core serializer assumes that the values on object fields are
correctly typed to their respective fields. Updates made to an object's
in-memory state may not meet this assumption. To prevent this violation, values
are typed by calling `to_python` from the field object, the result is set on a
copy of the instance and the copy is sent to the serializer.

For example, a datetime parsed as a string and set on a `models.DateTimeField()`
field is normally written to the database as a string and set to a python
datetime object when read on the model. Therefore, no error is raised on save
when a datetime field contains a string formatted datetime. The serializer,
however, expects a datetime object and will attempt to call `isoformat()` to
convert the datetime object to a string which then raises an attribute error.