Change strategy for skipping object fields
With this strategy, a value is not require to determine if the field
should be skipped when performing the ``to_python`` call. This is
preferred because there are a few edge cases around creating log
entries as a result of a cascade delete. In those cases, there are
situations when the system attempts to get a value that has already been
deleted.